pyproject.toml
[tool.poetry]
name = "yosemite-sampler"
version = "0.1.0"
description = ""
authors = ["Jeremy Gerard <jeremyegerard@gmail.com>"]
readme = "README.md"
packages = [{include = "yosemite_sampler"}]

[tool.poetry.dependencies]
python = "^3.11"
bs4 = "^0.0.1"
rich = "^13.4.2"
selenium = "^4.10.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
